Fix template type.

+++ src/PacketSerial.h
@@ -30,7 +30,7 @@
 #include "Encoding/SLIP.h"
 
 
-template<typename EncoderType, uint8_t PacketMarker = 0, int BufferSize = 256>
+template<typename EncoderType, uint8_t PacketMarker = 0, size_t BufferSize = 256>
 class PacketSerial_
 {
 public:
